shine: Don't assume int == long int
shine_encode_buffer expects an integer

+++ modules/codec/shine.c
@@ -246,7 +246,7 @@ static block_t *EncodeFrame( encoder_t *p_enc, block_t *p_block )
         aout_Deinterleave( pcm_planar_buf, p_pcm_block->p_buffer,
                 p_sys->samples_per_frame, p_enc->fmt_in.audio.i_channels, p_enc->fmt_in.i_codec);
 
-        long written;
+        int written;
         unsigned char *buf = shine_encode_buffer(p_sys->s, pcm_planar_buf_chans, &written);
         block_Release( p_pcm_block );
